// project-manager-web/src/main/web/src/app/model/task.ts
export class Task {
  taskId: number;
  task: string;
  startDate: Date;
  endDate: Date;
  priority: number;
  status: boolean;
  projectId: number;
  project: string;
  parentId: number;
  parentTask: string;
  userId: number;
  userName: string;
  userEmployeeId: number;
  startDateString : string;
  endDateString : string;
}